test(githubaction,jsii): added constructor options tests
